==> code/1.1.clustering_distance_matrix.ipynb <==
import pandas as pd
import re
import numpy as np
from scipy.spatial import distance

# Clustering Distance Matrix
Find Jaccard distances between all pain map areas.

Use R package DECIPHER to cluster using neighbor-joining algorithm.

pain_data = pd.read_csv('../data/painful_area_data.csv')
X_transposed = pain_data.transpose()

# Create a Jaccard distance matrix
areas = X_transposed.index
Jaccard_distance = pd.DataFrame(columns=areas, index=areas)
for a1 in areas:
    for a2 in areas:
        row1 = X_transposed.loc[X_transposed.index == a1]
        row2 = X_transposed.loc[X_transposed.index == a2]
        jd = distance.jaccard(row1, row2)
        
        Jaccard_distance.loc[Jaccard_distance.index == a2, a1] = jd
# Jaccard_distance.to_csv('../results/clustering/jaccard_distance_painful_areas.csv')
